requirements.txt
# Web服务框架
fastapi
uvicorn[standard]

# LangChain核心与Ollama集成
langchain
langchain-ollama

# 向量存储
faiss-cpu

# 环境配置
python-dotenv

# 文档加载与处理
langchain-community>=0.0.10
unstructured>=0.10.0
pypdf2>=3.0.0
python-docx>=0.8.11
unstructured[docx]>=0.10.0
unstructured[pdf]>=0.10.0

# LoRA微调相关依赖
transformers>=4.36.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
torch>=2.0.0
scikit-learn
numpy
scipy

# 测试相关依赖
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
requests>=2.28.0

# 调度器和文件监控依赖
schedule>=1.2.0
watchdog>=3.0.0
psutil>=5.9.0